add operator abstract class and adapted operator simulation
