+++ bulletroboy/cage_simulation.py
@@ -6,7 +6,7 @@ import rclpy
 from threading import Thread
 from rclpy.executors import MultiThreadedExecutor
 
-from bulletroboy.operator import Operator, Moves
+from bulletroboy.operator_simulation import OperatorSim, Moves
 from bulletroboy.exoforce import CageConfiguration
 from bulletroboy.exoforce_simulation import ExoForceSim
 
@@ -45,7 +45,7 @@ def main():
     initial_cage_conf = CageConfiguration(args.config_path)
 
     rclpy.init()
-    operator = Operator(human_model)
+    operator = OperatorSim(human_model)
     exoforce = ExoForceSim(initial_cage_conf, operator, args.mode)
 
     executor = MultiThreadedExecutor()
@@ -58,7 +58,7 @@ def main():
     # RUN SIM
     try:
         while True:
-            exoforce.operator.publish_state()
+            exoforce.operator.publish_ef_state()
             exoforce.move_operator_sim()
             exoforce.update()
             p.stepSimulation()
